Remove PlotDriverNoUnsure.py and main_000002.py, add generated chart image
