--- main_000002.py
@@ -1,130 +1,0 @@
-from Chan import CChan
-from ChanConfig import CChanConfig
-from Common.CEnum import AUTYPE, DATA_SRC, KL_TYPE
-from Plot.AnimatePlotDriver import CAnimateDriver
-from Plot.PlotDriverNoUnsure import CPlotDriver
-import matplotlib.pyplot as plt
-
-if __name__ == "__main__":
-    code = "000002.SZ"
-    begin_time = "2018-01-01"
-    end_time = None
-    data_src = DATA_SRC.BAO_STOCK
-    # 多级别K线：月线、周线、日线（从大级别到小级别）
-    lv_list = [KL_TYPE.K_MON, KL_TYPE.K_WEEK, KL_TYPE.K_DAY]
-
-    config = CChanConfig({
-        "bi_strict": True,
-        "trigger_step": False,
-        "skip_step": 0,
-        "divergence_rate": float("inf"),
-        "bsp2_follow_1": False,
-        "bsp3_follow_1": False,
-        "min_zs_cnt": 0,
-        "bs1_peak": False,
-        "macd_algo": "peak",
-        "bs_type": '1,2,3a,1p,2s,3b',
-        "print_warning": True,
-        "zs_algo": "normal",
-    })
-
-    # 为每个时间级别配置不同的绘图参数
-    plot_config = {
-        KL_TYPE.K_MON: {
-            "plot_kline": True,
-            "plot_kline_combine": True,
-            "plot_bi": True,
-            "plot_seg": True,
-            "plot_eigen": False,
-            "plot_zs": True,
-            "plot_macd": False,
-            "plot_mean": False,
-            "plot_channel": False,
-            "plot_bsp": True,
-            "plot_extrainfo": False,
-            "plot_demark": False,
-            "plot_marker": False,
-            "plot_rsi": False,
-            "plot_kdj": False,
-        },
-        KL_TYPE.K_WEEK: {
-            "plot_kline": True,
-            "plot_kline_combine": True,
-            "plot_bi": True,
-            "plot_seg": True,
-            "plot_eigen": False,
-            "plot_zs": True,
-            "plot_macd": True,
-            "plot_mean": False,
-            "plot_channel": False,
-            "plot_bsp": True,
-            "plot_extrainfo": False,
-            "plot_demark": False,
-            "plot_marker": False,
-            "plot_rsi": False,
-            "plot_kdj": False,
-        },
-        KL_TYPE.K_DAY: {
-            "plot_kline": True,
-            "plot_kline_combine": True,
-            "plot_bi": True,
-            "plot_seg": True,
-            "plot_eigen": False,
-            "plot_zs": True,
-            "plot_macd": True,
-            "plot_mean": False,
-            "plot_channel": False,
-            "plot_bsp": True,
-            "plot_extrainfo": False,
-            "plot_demark": False,
-            "plot_marker": False,
-            "plot_rsi": False,
-            "plot_kdj": False,
-        }
-    }
-
-    plot_para = {
-        "seg": {
-            # "plot_trendline": True,
-        },
-        "bi": {
-            # "show_num": True,
-            # "disp_end": True,
-        },
-        "figure": {
-            "x_range": 200,
-            "w": 24,
-            "h": 20,  # 增加高度以适应三个子图
-        },
-        "marker": {
-            # "markers": {  # text, position, color
-            #     '2023/06/01': ('marker here', 'up', 'red'),
-            #     '2023/06/08': ('marker here', 'down')
-            # },
-        }
-    }
-    
-    chan = CChan(
-        code=code,
-        begin_time=begin_time,
-        end_time=end_time,
-        data_src=data_src,
-        lv_list=lv_list,
-        config=config,
-        autype=AUTYPE.QFQ,
-    )
-
-    if not config.trigger_step:
-        plot_driver = CPlotDriver(
-            chan,
-            plot_config=plot_config,
-            plot_para=plot_para,
-        )
-        plot_driver.figure.show()
-        plot_driver.save2img("./000002_multi_timeframe_chart.png")
-    else:
-        CAnimateDriver(
-            chan,
-            plot_config=plot_config,
-            plot_para=plot_para,
-        )
